% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_network.R
\name{build_network}
\alias{build_network}
\title{Build a Network from Ego-Alter Edge Lists}
\usage{
build_network(x1, x2)
}
\arguments{
\item{x1}{A vector containing ego node IDs.}

\item{x2}{A vector containing alter node IDs.}
}
\value{
A list containing:
\describe{
  \item{net0}{An igraph network object.}
  \item{n}{Number of nodes in the network.}
  \item{data}{A data frame with node IDs and their degrees.}
}
}
\description{
This function constructs an undirected network using ego and alter IDs.
It removes isolated vertices, simplifies the graph by removing duplicate
edges and self-loops, and returns the resulting graph along with node
degree information.
}
\examples{
ego <- c(1,1,2,3,4,5)
alter <- c(2,3,3,4,5,1)
res <- build_network(ego, alter)
res

}
